Review comments for tree
