+++ cmd/tree/main.go
@@ -2,41 +2,11 @@ package main
 
 import (
 	"flag"
-	"fmt"
+	"log"
 	"os"
-	"strings"
+
+	internal "github.com/codescalersinternships/Coreutils-Rawan-Mostafa/internal"
 )
-
-func check(e error) {
-	if e != nil {
-		panic(e)
-	}
-}
-func formattedPrint(directoryPath string, depth int) {
-	fmt.Print("|")
-
-	for i := 0; i < depth; i++ {
-		fmt.Print("──")
-	}
-	fmt.Println(directoryPath)
-}
-func tree(directoryPath string, depth int, maxDepth int) {
-
-	items, err := os.ReadDir(directoryPath)
-	check(err)
-	pathParts := strings.Split(directoryPath, "/")
-	formattedPrint(pathParts[len(pathParts)-1], depth)
-
-	for _, item := range items {
-		if item.IsDir() && depth+1 <= maxDepth {
-			tree(directoryPath+"/"+item.Name(), depth+1, maxDepth)
-		} else if !item.IsDir() && depth+1 <= maxDepth {
-			formattedPrint(item.Name(), depth+1)
-		} else if depth+1 > maxDepth {
-			return
-		}
-	}
-}
 
 func main() {
 	var maxDepth int
@@ -44,9 +14,11 @@ func main() {
 	flag.Parse()
 	if len(flag.Args()) == 0 {
 		currDir, err := os.Getwd()
-		check(err)
-		tree(currDir, 0, maxDepth)
+		if err != nil {
+			log.Fatal("Error reaching the current directory")
+		}
+		internal.Tree(currDir, 0, maxDepth)
 	} else {
-		tree(flag.Args()[0], 0, maxDepth)
+		internal.Tree(flag.Args()[0], 0, maxDepth)
 	}
 }
